@@ -28,16 +29,14 @@ android {
     dataBinding {
         enabled true
     }
-    compileSdk 33
+    compileSdk 34
 
     defaultConfig {
         applicationId "jp.ceed.android.mylapslogger"
         minSdk 24
-        targetSdk 33
+        targetSdk 34
         versionCode Integer.parseInt(AppVersionCode)
         versionName AppVersionName
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         // Room database の設定
         javaCompileOptions {
