@@ -5,18 +5,18 @@ buildscript {
 		mavenCentral()
 	}
 
-	ext.kotlin_version = "1.8.0"
+	ext.kotlin_version = "1.9.23"
 	ext.koin_version = '3.1.4'
 
 	dependencies {
-		classpath 'com.android.tools.build:gradle:7.4.2'
+		classpath 'com.android.tools.build:gradle:8.3.2'
 		classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.8.0"
-		classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.5.3"
+		classpath "androidx.navigation:navigation-safe-args-gradle-plugin:2.7.7"
 		classpath "io.insert-koin:koin-gradle-plugin:$koin_version"
-		classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.5'
-		classpath 'com.google.dagger:hilt-android-gradle-plugin:2.45'
+		classpath 'com.google.firebase:firebase-crashlytics-gradle:2.9.9'
+		classpath 'com.google.dagger:hilt-android-gradle-plugin:2.48'
 		classpath "org.jetbrains.kotlin:kotlin-serialization:$kotlin_version"
-		classpath 'com.google.gms:google-services:4.3.15'
+		classpath 'com.google.gms:google-services:4.4.1'
 	}
 }
 
