@@ -4,15 +4,11 @@ import com.google.common.truth.Truth.assertThat
 import io.mockk.every
 import io.mockk.mockk
 import jp.ceed.android.mylapslogger.network.response.Sessions
-import org.junit.platform.runner.JUnitPlatform
-import org.junit.runner.RunWith
-import org.spekframework.spek2.Spek
-import org.spekframework.spek2.style.specification.describe
+import org.junit.jupiter.api.Test
 
-@RunWith(JUnitPlatform::class)
-object PracticeResultsRepositoryTest : Spek({
+class PracticeResultsRepositoryTest {
 
-    val sessions: Sessions = mockk {
+    private val sessions: Sessions = mockk {
         every {
             sessions
         } returns listOf( mockk {
@@ -31,15 +27,14 @@ object PracticeResultsRepositoryTest : Spek({
         })
     }
 
-    val practiceResultsRepository = PracticeResultsRepository(
+    private val practiceResultsRepository = PracticeResultsRepository(
         sessionsApiDataSource = mockk(),
         sessionDataCreator = mockk(),
     )
 
-    describe("次回ローディング時間の計算") {
-        it("計算結果が正しいこと") {
-            val result = practiceResultsRepository.calculateNextLoadTime(sessions)
-            assertThat(result).isEqualTo(1658042364093)
-        }
+    @Test
+    fun calculateNextLoadTime() {
+        val result = practiceResultsRepository.calculateNextLoadTime(sessions)
+        assertThat(result).isEqualTo(1658042364093)
     }
-})
+}
