@@ -7,17 +7,15 @@ import io.mockk.verify
 import jp.ceed.android.mylapslogger.initMainLooper
 import jp.ceed.android.mylapslogger.repository.AppInfoRepository
 import jp.ceed.android.mylapslogger.util.AppSettings
-import org.junit.platform.runner.JUnitPlatform
-import org.junit.runner.RunWith
-import org.spekframework.spek2.Spek
-import org.spekframework.spek2.style.specification.describe
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
 
-@RunWith(JUnitPlatform::class)
-object AppInfoFragmentViewModelTest : Spek({
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class AppInfoFragmentViewModelTest {
 
-    initMainLooper()
-
-    val appSettings: AppSettings = mockk(relaxed = true) {
+    private val appSettings: AppSettings = mockk(relaxed = true) {
         every {
             isShowPracticeResultsAsSeparate()
         } returns true
@@ -26,49 +24,55 @@ object AppInfoFragmentViewModelTest : Spek({
         } returns true
     }
 
-    val appInfoRepository: AppInfoRepository = mockk {
+    private val appInfoRepository: AppInfoRepository = mockk {
         every {
             getVersionName()
         } returns "1.2.3"
     }
 
-    val viewModel = AppInfoFragmentViewModel(
-        appSettings = appSettings,
-        appInfoRepository = appInfoRepository,
-    )
+    private lateinit var viewModel: AppInfoFragmentViewModel
 
-    describe("初期状態の確認"){
-        it("正しい値がセットされている"){
-            assertThat(viewModel.appVersionName.value).isEqualTo("1.2.3")
-            assertThat(viewModel.showPracticeResultAsSeparate.value).isEqualTo(true)
-            assertThat(viewModel.showSpeedBar.value).isEqualTo(true)
+    @BeforeAll
+    fun beforeAll() {
+        initMainLooper()
+    }
+
+    @BeforeEach
+    fun setUp() {
+        viewModel = AppInfoFragmentViewModel(
+            appSettings = appSettings,
+            appInfoRepository = appInfoRepository,
+        )
+    }
+
+    @Test
+    fun initialState() {
+        assertThat(viewModel.appVersionName.value).isEqualTo("1.2.3")
+        assertThat(viewModel.showPracticeResultAsSeparate.value).isEqualTo(true)
+        assertThat(viewModel.showSpeedBar.value).isEqualTo(true)
+    }
+
+    @Test
+    fun saveShowPracticeResultsAsSeparate() {
+        viewModel.saveShowPracticeResultsAsSeparate(true)
+        verify {
+            viewModel.appSettings.saveShowPracticeResultsAsSeparate(true)
         }
     }
 
-    describe("saveShowPracticeResultsAsSeparateが呼ばれたとき"){
-        it("保存の処理が呼ばれている"){
-            viewModel.saveShowPracticeResultsAsSeparate(true)
-            verify {
-                viewModel.appSettings.saveShowPracticeResultsAsSeparate(true)
-            }
+    @Test
+    fun saveShowSpeedBar() {
+        viewModel.saveShowSpeedBar(false)
+        verify {
+            viewModel.appSettings.saveSpeedBar(false)
         }
     }
 
-    describe("saveShowSpeedBarが呼ばれえた時"){
-        it("保存の処理が呼ばれている"){
-            viewModel.saveShowSpeedBar(false)
-            verify {
-                viewModel.appSettings.saveSpeedBar(false)
-            }
+    @Test
+    fun saveAllowSessionAutoLoading() {
+        viewModel.saveAllowSessionAutoLoading(false)
+        verify {
+            viewModel.appSettings.saveAllowSessionAutoLoading(false)
         }
     }
-
-    describe("saveAllowSessionAutoLoadingが呼ばれえた時"){
-        it("保存の処理が呼ばれている"){
-            viewModel.saveAllowSessionAutoLoading(false)
-            verify {
-                viewModel.appSettings.saveAllowSessionAutoLoading(false)
-            }
-        }
-    }
-})
+}
