@@ -6,35 +6,44 @@ import io.mockk.mockk
 import jp.ceed.android.mylapslogger.initMainLooper
 import jp.ceed.android.mylapslogger.model.FuelConsumptionListItem
 import jp.ceed.android.mylapslogger.repository.ActivityInfoTrackRepository
-import org.junit.platform.runner.JUnitPlatform
-import org.junit.runner.RunWith
-import org.spekframework.spek2.Spek
-import org.spekframework.spek2.style.specification.describe
+import org.junit.jupiter.api.BeforeAll
 
-@RunWith(JUnitPlatform::class)
-object FuelConsumptionListFragmentViewModelTest : Spek({
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
 
-    initMainLooper()
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class FuelConsumptionListFragmentViewModelTest {
 
-    val activityInfoTrackRepository: ActivityInfoTrackRepository = mockk {
+    private val activityInfoTrackRepository: ActivityInfoTrackRepository = mockk {
         coEvery {
             getFuelConsumptionList()
-        } returns listOf(FuelConsumptionListItem(
-            dateTime = "2022-01-01",
-            trackName = "trackName",
-            fuelConsumption = "1.2",
-        ))
+        } returns listOf(
+            FuelConsumptionListItem(
+                dateTime = "2022-01-01",
+                trackName = "trackName",
+                fuelConsumption = "1.2",
+            )
+        )
     }
 
-    val viewModel = FuelConsumptionListFragmentViewModel(
-        activityInfoTrackRepository
-    )
+    private lateinit var viewModel: FuelConsumptionListFragmentViewModel
 
-    describe("初期か処理の確認"){
-        it("listItemが正しく初期化されている事"){
-            assertThat(viewModel.listItem.value).hasSize(1)
-            assertThat(viewModel.listItem.value?.get(0)?.fuelConsumption).isEqualTo("1.2")
-        }
+    @BeforeAll
+    fun beforeAll() {
+        initMainLooper()
     }
 
-})
+    @BeforeEach
+    fun setUp() {
+        viewModel = FuelConsumptionListFragmentViewModel(
+            activityInfoTrackRepository
+        )
+    }
+
+    @Test
+    fun initialState() {
+        assertThat(viewModel.listItem.value).hasSize(1)
+        assertThat(viewModel.listItem.value?.get(0)?.fuelConsumption).isEqualTo("1.2")
+    }
+}
