@@ -8,43 +8,50 @@ import io.mockk.mockk
 import jp.ceed.android.mylapslogger.entity.PracticeTrack
 import jp.ceed.android.mylapslogger.initMainLooper
 import jp.ceed.android.mylapslogger.repository.PracticeTrackRepository
-import org.junit.platform.runner.JUnitPlatform
-import org.junit.runner.RunWith
-import org.spekframework.spek2.Spek
-import org.spekframework.spek2.style.specification.describe
+import org.junit.jupiter.api.BeforeAll
+import org.junit.jupiter.api.BeforeEach
+import org.junit.jupiter.api.Test
+import org.junit.jupiter.api.TestInstance
 
-@RunWith(JUnitPlatform::class)
-object PracticeByTrackFragmentViewModelTest : Spek({
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class PracticeByTrackFragmentViewModelTest {
 
-    initMainLooper()
-
-    val savedStateHandle: SavedStateHandle = mockk(relaxed = true) {
+    private val savedStateHandle: SavedStateHandle = mockk(relaxed = true) {
         every {
             get<Int>("trackId")
         } returns 123
     }
 
-    val listItem: PracticeTrack = mockk(relaxed = true) {
+    private val listItem: PracticeTrack = mockk(relaxed = true) {
         every {
             trackId
         } returns (savedStateHandle.get<Int>("trackId") ?: 0)
     }
 
-    val practiceTrackRepository: PracticeTrackRepository = mockk(relaxed = true) {
+    private val practiceTrackRepository: PracticeTrackRepository = mockk(relaxed = true) {
         coEvery {
             getPracticeListByTrack(any(), true)
         } returns listOf(listItem)
     }
 
-    val viewModel = PracticeByTrackFragmentViewModel(
-        savedStateHandle = savedStateHandle,
-        practiceTrackRepository = practiceTrackRepository,
-    )
+    private lateinit var viewModel: PracticeByTrackFragmentViewModel
 
-    describe("初期値の確認") {
-        it("フィールドに正しい値がセットされている") {
-            assertThat(viewModel.practiceTrackList.value).hasSize(1)
-            assertThat(viewModel.practiceTrackList.value?.get(0)?.trackId).isEqualTo(123)
-        }
+    @BeforeAll
+    fun beforeAll() {
+        initMainLooper()
     }
-})
+
+    @BeforeEach
+    fun setUp() {
+        viewModel = PracticeByTrackFragmentViewModel(
+            savedStateHandle = savedStateHandle,
+            practiceTrackRepository = practiceTrackRepository,
+        )
+    }
+
+    @Test
+    fun initState() {
+        assertThat(viewModel.practiceTrackList.value).hasSize(1)
+        assertThat(viewModel.practiceTrackList.value?.get(0)?.trackId).isEqualTo(123)
+    }
+}
